+++ packages/preview/src/plugins/remote-plugin-core/remotePluginCore.ts
@@ -1,0 +1,3 @@
+import { RemotePlugin } from '../remotePluginApi'
+
+export const remotePluginCore: RemotePlugin = () => {}