+++ packages/preview/src/plugins/remote-plugin-render/remotePluginRender.ts
@@ -1,0 +1,8 @@
+import { RemotePlugin } from '../remotePluginApi'
+
+export const remotePluginRender: RemotePlugin = api => {
+  api.messageChannel.onMessage('updateRender', payload => {
+    api.component.render = new Function(payload.render)
+    api.component.scheduleRerender()
+  })
+}