+++ packages/preview/src/plugins/remote-plugin-style/remotePluginStyle.ts
@@ -1,0 +1,9 @@
+import { RemotePlugin } from '../remotePluginApi'
+
+const $style = document.getElementById('style') as HTMLStyleElement
+
+export const remotePluginStyle: RemotePlugin = api => {
+  api.messageChannel.onMessage('updateStyle', payload => {
+    $style.innerText = payload.style
+  })
+}