+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 | Version | Date     | Changelog                                       |
 | ------- | -------- | ----------------------------------------------- |
+| 1.2.0   | 16.10.19 | &mdash; **[Feature]** Preview Props             |
+| 1.1.1   | 16.10.19 | &mdash; **[Fix]** Don't open by default         |
 | 1.1.0   | 09.10.19 | &mdash; **[Feature]** Open Preview with Command |
 | 1.0.1   | 09.10.19 | &mdash; First working release                   |
 | 1.0.0   | 09.10.19 | &mdash; First release                           |